requirements.txt
dash==4.1.0
dash_ag_grid==35.2.0
dash_bio==1.0.2
biopython==1.87
pandas==3.0.2
numpy==2.4.4